OS_기초/requirements.txt
fastapi
uvicorn[standard]
azure-cognitiveservices-speech
python-dotenv
Werkzeug
Jinja2